Simplified DSL
Now music is the top level production and includes voices (wiich introduces parallel lines).  This means that all constructors are amenable to the Modify statement.  However, it also means that there is an option to parallelise a melody to an arbitrary depth.

Probably no bad thing if we later adopt a policy of defaulting each parallel line to channel 0 or else requiring that the user selects an instrument/channel using the Modify command.