Allow Seq to have control as a child
This is necessary for the expressibility of the DSL - you want children of Seq to be able to modify individual lines.  The expense is that we have to surround the music introduced by each control modification to be surrounded by brackets (that experiment seems to have worked).  Updated README in line with this.