Added 'replece soundfonts' button
This allows you to replace the default set with another set that you have chosen using the dropdown menu.  You then have to change the text of the tune so that you use these instruments.

Mostly working.  For some reason, sometimes an instrument fails to be identified properly with a channel.

Added Frere Jacques as the default tune.  (You need to edit this once at the moment for it to be parsed)