Add basic support for ABC polyphony in editor
The ABC module now supports polyphonic ABC tunes.  This allows us to
import the ABC voices as separate PSoM DSL voices which are joined at
the top level by a Par construct.

Almost working.  We just have to sort out the DSL program name (which is
repeated in each voice and shouldn't be).